Use python3 binary to run unoconv
python version may differ in different distributions but it is common
practice to provide python3. This binary will also be local and not overwritten by pyenv or locally
in by pyenv so tests can be run locally again.

This fixes issue CI currently not running as docker container updated to
Debian buster.