fix!: use "Listing" instead of "RichText" for multiline support
BREAKING CHANGE: It turns out that docx-template already features a "Listing" class that
is a much better fit for our use-case (multiline-support, preserved
formatting, no changes in templates) than "RichText".

The listing class only handles \n, \a, \t and \f, see
https://github.com/elapouya/python-docx-template/blob/4fe1cc507d23cca68a8d78208dcb803a312578ec/docxtpl/__init__.py#L701

This wraps all values as "Listing" and adds a string cast to the jinja
filters for compatibility.