fix(multiline): revert automatic conversion, add filter instead
BREAKING CHANGE: After gathering some practical experience with the new
automatic "Listing"-conversion for multiline we noticed that this
feature is a little bit too "clever" and breaks many advanced use-cases.

This reverts the automatic conversion in favor of a simple "multiline"
filter that wraps the value in a "Listing".