feat(validation): validate docx template on upload
Docx templates can contain syntax errors that will only get caught when
rendering the template.  This causes problems, as the user seeing the
problem (some frontend user for example) may not be the same person who
caused the problem (the person creating the template).

Thus we now validate the syntax on upload already, as far as we can by
trying to render the templates on upload (works for both engines).

* Validation for the mailmerge engine ensures that the template has
  *some* placeholders.
* Validation for the docx template engine validates the syntax as well.