feat(validation): check template for available placeholders
This introduces the functionality that you can now pass a list of
available placeholders when creating a template. The template is then
checked to see whether it contains a strict subset of the given placeholders.

The syntax is a bit arbitrary, but allows specification of nested values,
and even lists:

    {
        "slug": ...,
        "available_placeholders": [
            "foo", "bar",
            "a_list[].property",
            "a_nested.object.with.property"
        ]
    }