feat(docx-template): support inline images for docx-template
This commit implements inline images for docx-templates. This adds a new
`image` jinja filter. When passed a image filename, it will
include the provided file with the same name in the document as inline image.

The `merge` serializer now accepts multipart requests with an optional list of
image files.

Normal json requests without images still work as expected.