fix: move temporary path to data directory
Since we hide /tmp from LibreOffice processes we can't use /tmp for temporary
files. I first used $MEDIA_ROOT/__convert__, but this directory is not an
internal directory as I expected (it was a bad idea), so I moved __convert__ to
/var/lib/document_merge_service/data/tmp ($DATABASE_DIR/tmp). The Dockerfile
will set the rights correctly: shared between the dms-user and root.