fix: delete old file if new file is being uploaded
Previously a new file with the same name would later replace the old
file, and no cleanup would thus be necessary; comparing the file name
thus made sense.

Since Django will now create the file with a random suffix to avoid file
conflicts and accidental clobbering, this approach no longer works, and
old files thus remain in the storage unreferenced.

The new approach is to check if a new file is being uploaded, and if so,
we delete the old file if there is any.