refactor: migrate to dedicated unoserver container
The old `unoconv` is not supported anymore upstream and needs to
be replaced. Instead of running all the subprocess management code
ourselves, we now just let this happen in it's own container. This
should improve service stability significantly, as it's now cleanly
separated.

While we're at it, also update the base image to newest debian Trixie